[fix] Fix 3 critical client bugs: FEC reconstruction, timewarp shader, shard truncation
- Implement Reed-Solomon FEC reconstruction in fec_decoder.cpp using
  GF(2^8) Vandermonde matrix inversion matching reed-solomon-erasure crate.
  Previously filled missing shards with zeros (no error correction).
- Fix timewarp fragment shader: sampler2D → samplerExternalOES to match
  GL_TEXTURE_EXTERNAL_OES from MediaCodec. Was producing black screen.
- Widen shard count types from uint8_t to uint16_t in FecFrameDecoder
  to support IDR frames exceeding 255 shards at high bitrates.

+++ client/app/src/main/cpp/fec_decoder.h
@@ -21,10 +21,10 @@ public:
     };
 
     /// Start collecting shards for a new frame.
-    void beginFrame(uint32_t frameIndex, uint8_t totalShards, uint8_t dataShards, bool isKeyframe);
+    void beginFrame(uint32_t frameIndex, uint16_t totalShards, uint16_t dataShards, bool isKeyframe);
 
     /// Add a received shard.
-    void addShard(uint8_t shardIndex, const uint8_t* data, int dataLen);
+    void addShard(uint16_t shardIndex, const uint8_t* data, int dataLen);
 
     /// Attempt to reconstruct the frame.
     /// Returns the decoded frame data if enough shards are available.
@@ -37,11 +37,11 @@ public:
 
 private:
     uint32_t m_frameIndex = 0;
-    uint8_t m_totalShards = 0;
-    uint8_t m_dataShards = 0;
+    uint16_t m_totalShards = 0;
+    uint16_t m_dataShards = 0;
     bool m_isKeyframe = false;
     int m_shardSize = 0;
-    uint8_t m_receivedCount = 0;
+    uint16_t m_receivedCount = 0;
 
     // Shard storage: index → data (empty = not received)
     std::vector<std::vector<uint8_t>> m_shards;
